// src/app/core/api/generated/model/patientUpsertRequestDto.ts
/**
 * Patient Account Tool API
 *
 * Contact: platform@example.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PatientUpsertRequestDto { 
    accountNumber: string;
    mrn?: string;
    firstName: string;
    middleName?: string;
    lastName: string;
    dateOfBirth: string;
    gender: string;
    phoneNumber?: string;
    email?: string;
    ssnLast4?: string;
    status?: string;
    statusReason?: string;
}